- Add `isSelecting` address - Update `getSelection` logics - Use int64_t in struct `CaretDimension`

+++ types/CaretDimension.h
@@ -4,6 +4,6 @@
 
 namespace types {
     struct CaretDimension {
-        uint32_t height, xPosition, yPosition;
+        int64_t height, xPosition, yPosition;
     };
 }
